Refactor

+++ Graphics3D/Quaternion.cs
@@ -8,13 +8,13 @@ namespace Graphics3D
 {
     class Quaternion
     {
-        public double X;
-        public double Y;
-        public double Z;
-        public double W;
+        public float X;
+        public float Y;
+        public float Z;
+        public float W;
 
         public Quaternion() { }
-        public Quaternion(double x, double y, double z, double w)
+        public Quaternion(float x, float y, float z, float w)
         {
             X = x;
             Y = y;
@@ -22,18 +22,18 @@ namespace Graphics3D
             W = w;
         }
 
-        public static Quaternion RotationYawPitchRoll(double yaw, double pitch, double roll)
+        public static Quaternion RotationYawPitchRoll(float yaw, float pitch, float roll)
         {
-            double halfRoll = roll * 0.5f;
-            double halfPitch = pitch * 0.5f;
-            double halfYaw = yaw * 0.5f;
+            float halfRoll = roll * 0.5f;
+            float halfPitch = pitch * 0.5f;
+            float halfYaw = yaw * 0.5f;
 
-            double sinRoll = Math.Sin(halfRoll);
-            double cosRoll = Math.Cos(halfRoll);
-            double sinPitch = Math.Sin(halfPitch);
-            double cosPitch = Math.Cos(halfPitch);
-            double sinYaw = Math.Sin(halfYaw);
-            double cosYaw = Math.Cos(halfYaw);
+            float sinRoll = (float)Math.Sin(halfRoll);
+            float cosRoll = (float)Math.Cos(halfRoll);
+            float sinPitch = (float)Math.Sin(halfPitch);
+            float cosPitch = (float)Math.Cos(halfPitch);
+            float sinYaw = (float)Math.Sin(halfYaw);
+            float cosYaw = (float)Math.Cos(halfYaw);
 
             Quaternion result = new Quaternion();
             result.X = (cosYaw * sinPitch * cosRoll) + (sinYaw * cosPitch * sinRoll);
